refactor/ nav destructured in various components
